Update pkidestroy and pki-server remove
pkidestroy and pki-server remove commands have been modified to
work more consistently. Now by default the logs will be retained
so the following folders will remain:

- /var/lib/pki/<instance>
- /var/lib/pki/<instance>/logs
- /var/log/pki/<instance> for instances using FHS layout

The tools will also provide an option to remove the logs if they
are no longer needed.

The tests have been updated accordingly.

+++ base/server/python/pki/server/instance.py
@@ -394,18 +394,19 @@ class PKIInstance(pki.server.PKIServer):
         for external_cert in PKIInstance.read_external_certs(conf_file):
             self.external_certs.append(external_cert)
 
-    def remove(self, force=False):
+    def remove(self, remove_logs=False, force=False):
 
         logger.info('Removing %s', self.unit_file)
         pki.util.unlink(self.unit_file, force=force)
 
         self.remove_registry(force=force)
 
-        logs_link = os.path.join(self.base_dir, 'logs')
-        logger.info('Removing %s', logs_link)
-        pki.util.unlink(logs_link, force=force)
+        if remove_logs:
+            logs_link = os.path.join(self.base_dir, 'logs')
+            logger.info('Removing %s', logs_link)
+            pki.util.unlink(logs_link, force=force)
 
-        super().remove(force=force)
+        super().remove(remove_logs=remove_logs, force=force)
 
     def remove_libs(self, force=False):
